+++ packages/analytics-browser-utils/src/cookie-name.ts
