+++ packages/analytics-browser-utils/test/attribution/campaign-parser.test.ts
@@ -1,6 +1,6 @@
 import { BASE_CAMPAIGN } from '../../src/attribution/constants';
 import { CampaignParser } from '../../src/attribution/campaign-parser';
-import * as queryParams from '../../src/utils/query-params';
+import * as queryParams from '../../src/query-params';
 
 describe('campaign-parser', () => {
   describe('parse', () => {
