+++ packages/analytics-browser-utils/test/helpers/constants.ts
@@ -1,0 +1,1 @@
+export const API_KEY = 'apiKey';
