+++ packages/analytics-react-native/package.json
@@ -57,6 +57,7 @@
     "url": "https://github.com/amplitude/Amplitude-TypeScript/issues"
   },
   "dependencies": {
+    "@amplitude/analytics-browser-utils": "^0.0.0",
     "@amplitude/analytics-connector": "1.4.5",
     "@amplitude/analytics-core": "^0.8.0",
     "@amplitude/analytics-types": "^0.8.0",
