+++ packages/analytics-react-native/src/config.ts
@@ -8,13 +8,15 @@ import {
   SessionManager as ISessionManager,
 } from '@amplitude/analytics-types';
 import { Config, MemoryStorage, UUID } from '@amplitude/analytics-core';
+import {
+  CookieStorage,
+  getCookieName,
+  getQueryParams,
+  SessionManager,
+  FetchTransport,
+} from '@amplitude/analytics-browser-utils';
 
-import { CookieStorage } from './storage/cookie';
-import { FetchTransport } from './transports/fetch';
 import { LocalStorage } from './storage/local-storage';
-import { getCookieName } from './utils/cookie-name';
-import { getQueryParams } from './utils/query-params';
-import { SessionManager } from './session-manager';
 
 export const getDefaultConfig = () => {
   const cookieStorage = new MemoryStorage<UserSession>();
