+++ packages/analytics-react-native/src/cookie-migration/index.ts
@@ -1,7 +1,6 @@
 import { BrowserOptions, Storage, UserSession } from '@amplitude/analytics-types';
-import { getOldCookieName } from '../utils/cookie-name';
+import { getOldCookieName, CookieStorage } from '@amplitude/analytics-browser-utils';
 import { LocalStorage } from '../storage/local-storage';
-import { CookieStorage } from '../storage/cookie';
 
 export const parseOldCookies = async (apiKey: string, options?: BrowserOptions): Promise<UserSession> => {
   let storage: Storage<string> = new CookieStorage<string>();
